+++ proptest/src/test_runner/config.rs
@@ -1,5 +1,5 @@
 //-
-// Copyright 2017, 2018 The proptest developers
+// Copyright 2017, 2018, 2019 The proptest developers
 //
 // Licensed under the Apache License, Version 2.0 <LICENSE-APACHE or
 // http://www.apache.org/licenses/LICENSE-2.0> or the MIT license
@@ -291,7 +291,13 @@ pub struct Config {
     /// Give up on shrinking if more than this number of iterations of the test
     /// code are run.
     ///
+    /// Setting this to `std::u32::MAX` causes the actual limit to be four
+    /// times the number of test cases.
+    ///
     /// Setting this value to `0` disables shrinking altogether.
+    ///
+    /// Note that the type of this field will change in a future version of
+    /// proptest to better accommodate its special values.
     ///
     /// The default is `std::u32::MAX`, which can be overridden by setting the
     /// `PROPTEST_MAX_SHRINK_ITERS` environment variable.
@@ -447,6 +453,17 @@ impl Config {
     #[cfg(not(feature = "timeout"))]
     pub fn timeout(&self) -> u32 {
         0
+    }
+
+    /// Returns the configured limit on shrinking iterations.
+    ///
+    /// This takes into account the special "automatic" behaviour.
+    pub fn max_shrink_iters(&self) -> u32 {
+        if u32::MAX == self.max_shrink_iters {
+            self.cases.saturating_mul(4)
+        } else {
+            self.max_shrink_iters
+        }
     }
 
     // Used by macros to force the config to be owned without depending on